llm-eval-lab/foundations/prompt-regression/requirements.txt
google-generativeai==0.8.3
openai==1.58.1
pytest==8.3.4
python-dotenv==1.0.1
PyYAML==6.0.2
